refactor: split backups helper view models
