+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - [VS-1765] UI repository creation and selected fire-and-forget work now use shared factory and detached-task helpers.
 - [VS-1766] Projects and Settings helper view models now live in focused files, with shared export-folder opening.
 - [VS-1767] Metadata tombstone exports and Backups option setters now reuse focused helper plumbing.
+- [VS-1768] Backups helper view models now live in a focused models file to shrink the main Backups view model.
 ### Fixed
 - [BUG-18038] Windows notification failures no longer assign an invalid empty toast tag and repeated OS-toast failures are suppressed after the first diagnostic entry.
 - [BUG-18039] Manual storage-health rechecks now marshal notification and tray updates back to the UI thread after background drive probing.
