+++ ROADMAP.md
@@ -886,6 +886,9 @@
 - [x] `VS-1767` `P2` Consolidate tombstone export and Backups option helper plumbing. _(Done; tracked by #340)_
   - Scope: share metadata tombstone meta-info stamping, tombstone row creation, and Backups silent option setter logic.
   - Acceptance: tombstone exporters keep existing writer metadata and lock-retry behavior; Backups option selections keep current transient-null handling.
+- [x] `VS-1768` `P2` Split Backups helper view models into a focused file. _(Done; tracked by #345)_
+  - Scope: move Backups helper model/view-model classes out of the oversized main Backups view model while preserving binding-compatible names and namespaces.
+  - Acceptance: the main Backups view model is easier to scan; helper classes retain existing behavior; solution build remains clean.
 
 ### Proposed delivery phases
 1. Phase `A` (integrity backbone): `VS-1706` -> `VS-1711` -> `VS-1701` -> `VS-1705` -> `VS-1712`
