+++ src/VaultSync.UI/ViewModels/BackupsViewModel.cs
@@ -3874,1210 +3874,4 @@ namespace VaultSync.UI.ViewModels
             int Changes,
             long ChangedBytes);
     }
-
-    // ---------- Models ----------
-
-        public sealed class StorageConsumerItem
-        {
-            public StorageConsumerItem(string projectName, string totalSize, double sharePercent)
-            {
-                ProjectName = projectName;
-                TotalSize = totalSize;
-                SharePercent = sharePercent;
-            }
-
-            public string ProjectName { get; }
-            public string TotalSize { get; }
-            public double SharePercent { get; }
-            public string SharePercentLabel => $"{SharePercent:0}%";
-        }
-
-        public sealed class DiffPreviewPathItem
-        {
-            public DiffPreviewPathItem(string path, int changes, string changedBytes)
-            {
-                Path = path;
-                Changes = changes;
-                ChangedBytes = changedBytes;
-            }
-
-            public string Path { get; }
-            public int Changes { get; }
-            public string ChangedBytes { get; }
-        }
-
-        public class BackupSnapshotItem : ViewModelBase
-        {
-        public string Id { get; set; } = string.Empty;
-        public DateTime Timestamp { get; set; }
-        public long SizeBytes { get; set; }
-        private bool _isProtected;
-
-        /// <summary>Run trigger type, e.g. "Auto" or "Manual".</summary>
-        public string Type { get; set; } = "Manual";
-
-        /// <summary>Status, e.g. "Completed", "Failed".</summary>
-        public string Status { get; set; } = "Completed";
-
-        /// <summary>Label shown inside the tag pill.</summary>
-        public string? Label { get; set; }
-
-        /// <summary>Localized backup mode label for display (Full/Incremental/Imported context).</summary>
-        public string TypeLabel { get; set; } = string.Empty;
-        public string ModeChipLabel { get; set; } = string.Empty;
-        public string EncryptionChipLabel { get; set; } = string.Empty;
-        public string RetentionDefaultLabel { get; set; } = "Retention: eligible for pruning";
-        public string RetentionProtectedLabel { get; set; } = "Retention: kept (protected)";
-        public string RetentionOutcomeLabel => IsProtected ? RetentionProtectedLabel : RetentionDefaultLabel;
-
-        /// <summary>Optional project id this snapshot belongs to; null for global.</summary>
-        public string? ProjectId { get; set; }
-
-        /// <summary>Destination endpoint that stored this backup.</summary>
-        public string DestinationDisplay { get; set; } = string.Empty;
-        public string DiffSummaryDisplay { get; set; } = string.Empty;
-        public string DiffTopPathsDisplay { get; set; } = string.Empty;
-        public bool HasDiffTopPaths { get; set; }
-        public bool CanOpenDiffDetails { get; set; }
-        public int DiffAdded { get; set; }
-        public int DiffModified { get; set; }
-        public int DiffDeleted { get; set; }
-        public long DiffNetBytes { get; set; }
-        public string DiffTopPathsJson { get; set; } = "[]";
-
-        public string SizeFormatted => FormatSize(SizeBytes);
-        public string TimelineSelectionLabel =>
-            $"{Timestamp:yyyy-MM-dd HH:mm} \u00b7 {TypeLabel} \u00b7 {SizeFormatted}";
-
-        public bool IsImported { get; set; }
-        public bool IsEncrypted { get; set; }
-
-        /// <summary>Localized label for the imported tag.</summary>
-        public string ImportedLabel { get; set; } = string.Empty;
-        public string EncryptionLabel { get; set; } = string.Empty;
-        public string OriginMachineName { get; set; } = string.Empty;
-
-        public bool IsProtected
-        {
-            get => _isProtected;
-            set
-            {
-                if (_isProtected != value)
-                {
-                    _isProtected = value;
-                    OnPropertyChanged(nameof(IsProtected));
-                    OnPropertyChanged(nameof(RetentionOutcomeLabel));
-                }
-            }
-        }
-
-        // ---------- Tag pill background color ----------
-
-        private static readonly IBrush DefaultBrush =
-            new ImmutableSolidColorBrush(Color.Parse("#22FFFFFF"));
-
-        // Auto snapshots: blue-ish
-        private static readonly IBrush AutoBrush =
-            new ImmutableSolidColorBrush(Color.Parse("#333A7AFE"));
-
-        // Manual snapshots: purple-ish
-        private static readonly IBrush ManualBrush =
-            new ImmutableSolidColorBrush(Color.Parse("#334568F2"));
-
-        // Imported snapshots: teal-ish
-        private static readonly IBrush ImportedBrush =
-            new ImmutableSolidColorBrush(Color.Parse("#3346C6A1"));
-
-        // Failed snapshots: red-ish
-        private static readonly IBrush FailedBrush =
-            new ImmutableSolidColorBrush(Color.Parse("#33FF4B4B"));
-
-        public IBrush TagBackground
-        {
-            get
-            {
-                if (string.Equals(Status, "Failed", StringComparison.OrdinalIgnoreCase))
-                    return FailedBrush;
-
-                if (string.Equals(Type, "Auto", StringComparison.OrdinalIgnoreCase))
-                    return AutoBrush;
-
-                if (string.Equals(Type, "Manual", StringComparison.OrdinalIgnoreCase))
-                    return ManualBrush;
-
-                return DefaultBrush;
-            }
-        }
-
-        public static IBrush ImportedTagBackground => ImportedBrush;
-
-        internal static string FormatSize(long bytes) =>
-            UiFormat.FormatBytes(bytes);
-    }
-
-    public class SnapshotProjectGroup
-    {
-        public string? ProjectId { get; set; }
-        public string ProjectName { get; set; } = string.Empty;
-        public string ProjectTagsDisplay { get; set; } = string.Empty;
-        public ObservableCollection<ProjectTagChip> ProjectTagChips { get; } = [];
-        public bool HasProjectTags => ProjectTagChips.Count > 0;
-        public string Summary { get; set; } = string.Empty;
-        public string TotalSizeFormatted { get; set; } = string.Empty;
-        public string LatestBackupDisplay { get; set; } = string.Empty;
-        public IBrush AccentBrush { get; set; } = new ImmutableSolidColorBrush(Color.Parse("#33405A"));
-        public bool IsExpanded { get; set; }
-
-        public ObservableCollection<BackupSnapshotItem> Snapshots { get; } =
-            [];
-    }
-
-    public sealed class BackupsProjectSortOption
-    {
-        public BackupsProjectSortOption(string id, string label)
-        {
-            Id = id;
-            Label = label;
-        }
-
-        public string Id { get; }
-        public string Label { get; }
-
-        public override string ToString() => Label;
-    }
-
-    public class ProjectBackupItem : ViewModelBase
-    {
-        public string Id { get; set; } = string.Empty;
-        public string Name { get; set; } = string.Empty;
-        public string ExternalId { get; set; } = string.Empty;
-        private string _projectTagsCsv = string.Empty;
-        public string ProjectTagsCsv
-        {
-            get => _projectTagsCsv;
-            set
-            {
-                if (!SetField(ref _projectTagsCsv, value ?? string.Empty, nameof(ProjectTagsCsv)))
-                    return;
-
-                RebuildProjectTags();
-                OnPropertyChanged(nameof(HasProjectTags));
-                OnPropertyChanged(nameof(ProjectTagsDisplay));
-                OnPropertyChanged(nameof(PrimaryTagSortKey));
-            }
-        }
-
-        public ObservableCollection<ProjectTagChip> ProjectTagChips { get; } = [];
-        public bool HasProjectTags => ProjectTagChips.Count > 0;
-        public string ProjectTagsDisplay => string.Join(", ", ProjectTagChips.Select(tag => tag.Value));
-        public string PrimaryTagSortKey => ProjectTagChips.FirstOrDefault()?.Value ?? string.Empty;
-
-        public DateTime? LastBackupTime { get; set; }
-        public int       SnapshotCount  { get; set; }
-        public long      TotalSizeBytes { get; set; }
-        public long?     StorageDeltaBytes { get; set; }
-        private string _restoreReadinessLabel = string.Empty;
-        public string RestoreReadinessLabel
-        {
-            get => _restoreReadinessLabel;
-            set => SetField(ref _restoreReadinessLabel, value ?? string.Empty, nameof(RestoreReadinessLabel));
-        }
-
-        private string _restoreReadinessReason = string.Empty;
-        public string RestoreReadinessReason
-        {
-            get => _restoreReadinessReason;
-            set => SetField(ref _restoreReadinessReason, value ?? string.Empty, nameof(RestoreReadinessReason));
-        }
-
-        private IBrush _restoreReadinessBrush = new ImmutableSolidColorBrush(Color.Parse("#7F8FA8"));
-        public IBrush RestoreReadinessBrush
-        {
-            get => _restoreReadinessBrush;
-            set => SetField(ref _restoreReadinessBrush, value, nameof(RestoreReadinessBrush));
-        }
-
-        public bool AutoBackupEnabled
-        {
-            get => _autoBackupEnabled;
-            set
-            {
-                if (!SetField(ref _autoBackupEnabled, value, nameof(AutoBackupEnabled)))
-                    return;
-                AutoBackupChanged?.Invoke(this);
-            }
-        }
-
-        private bool _autoBackupEnabled = true;
-        public Action<ProjectBackupItem>? AutoBackupChanged { get; set; }
-        public Action<ProjectBackupItem>? PreferredDestinationChanged { get; set; }
-        public Action<ProjectBackupItem>? EncryptionPolicyChanged { get; set; }
-        public Action<ProjectBackupItem>? RestoreModeChanged { get; set; }
-        public Action<ProjectBackupItem>? VerificationPolicyChanged { get; set; }
-
-        private string _preferredDestinationId = string.Empty;
-        public string PreferredDestinationId
-        {
-            get => _preferredDestinationId;
-            set => SetField(ref _preferredDestinationId, value ?? string.Empty, nameof(PreferredDestinationId));
-        }
-
-        private DestinationOption? _preferredDestinationOption;
-        public DestinationOption? PreferredDestinationOption
-        {
-            get => _preferredDestinationOption;
-            set
-            {
-                // Ignore transient null selection events fired while destination options refresh.
-                // Real "Auto" selection is represented by a non-null option with empty Id.
-                if (value is null)
-                    return;
-
-                if (ReferenceEquals(_preferredDestinationOption, value))
-                    return;
-
-                string previousId = _preferredDestinationOption?.Id ?? string.Empty;
-                _preferredDestinationOption = value;
-                OnPropertyChanged(nameof(PreferredDestinationOption));
-
-                string nextId = value.Id ?? string.Empty;
-                if (string.Equals(previousId, nextId, StringComparison.OrdinalIgnoreCase))
-                    return;
-
-                PreferredDestinationId = nextId;
-                PreferredDestinationChanged?.Invoke(this);
-            }
-        }
-
-        private string _preferredDestinationDisplay = string.Empty;
-        public string PreferredDestinationDisplay
-        {
-            get => _preferredDestinationDisplay;
-            set => SetField(ref _preferredDestinationDisplay, value ?? string.Empty, nameof(PreferredDestinationDisplay));
-        }
-
-        public void SetPreferredDestinationOption(DestinationOption? option)
-        {
-            SetOptionSilently(ref _preferredDestinationOption, option, nameof(PreferredDestinationOption));
-        }
-
-        private string _encryptionPolicy = ProjectEncryptionPolicy.Inherit;
-        public string EncryptionPolicy
-        {
-            get => _encryptionPolicy;
-            set => SetField(ref _encryptionPolicy, ProjectEncryptionPolicy.Normalize(value), nameof(EncryptionPolicy));
-        }
-
-        private string _encryptionKeyRef = string.Empty;
-        public string EncryptionKeyRef
-        {
-            get => _encryptionKeyRef;
-            set => SetField(ref _encryptionKeyRef, value ?? string.Empty, nameof(EncryptionKeyRef));
-        }
-
-        private EncryptionPolicyOption? _encryptionPolicyOption;
-        public EncryptionPolicyOption? EncryptionPolicyOption
-        {
-            get => _encryptionPolicyOption;
-            set
-            {
-                if (!SetField(ref _encryptionPolicyOption, value, nameof(EncryptionPolicyOption)))
-                    return;
-
-                // Ignore transient null selection events fired while option sources refresh.
-                // Real "inherit" selection is represented by a non-null option with Id="inherit".
-                if (value is null)
-                    return;
-
-                EncryptionPolicy = value.Id;
-                EncryptionPolicyChanged?.Invoke(this);
-            }
-        }
-
-        public void SetEncryptionPolicyOption(EncryptionPolicyOption? option)
-        {
-            SetOptionSilently(ref _encryptionPolicyOption, option, nameof(EncryptionPolicyOption));
-        }
-
-        private string _restoreMode = ProjectRestoreMode.Direct;
-        public string RestoreMode
-        {
-            get => _restoreMode;
-            set => SetField(ref _restoreMode, ProjectRestoreMode.Normalize(value), nameof(RestoreMode));
-        }
-
-        private RestoreModeOption? _restoreModeOption;
-        public RestoreModeOption? RestoreModeOption
-        {
-            get => _restoreModeOption;
-            set
-            {
-                if (!SetField(ref _restoreModeOption, value, nameof(RestoreModeOption)))
-                    return;
-
-                if (value is null)
-                    return;
-
-                RestoreMode = value.Id;
-                RestoreModeChanged?.Invoke(this);
-            }
-        }
-
-        public void SetRestoreModeOption(RestoreModeOption? option)
-        {
-            SetOptionSilently(ref _restoreModeOption, option, nameof(RestoreModeOption));
-        }
-
-        private string _verificationPolicy = ProjectVerificationPolicy.Always;
-        public string VerificationPolicy
-        {
-            get => _verificationPolicy;
-            set => SetField(ref _verificationPolicy, ProjectVerificationPolicy.Normalize(value), nameof(VerificationPolicy));
-        }
-
-        private VerificationPolicyOption? _verificationPolicyOption;
-        public VerificationPolicyOption? VerificationPolicyOption
-        {
-            get => _verificationPolicyOption;
-            set
-            {
-                if (!SetField(ref _verificationPolicyOption, value, nameof(VerificationPolicyOption)))
-                    return;
-
-                if (value is null)
-                    return;
-
-                VerificationPolicy = value.Id;
-                VerificationPolicyChanged?.Invoke(this);
-            }
-        }
-
-        public void SetVerificationPolicyOption(VerificationPolicyOption? option)
-        {
-            SetOptionSilently(ref _verificationPolicyOption, option, nameof(VerificationPolicyOption));
-        }
-
-        private void SetOptionSilently<T>(ref T? field, T? option, string propertyName)
-            where T : class
-        {
-            if (ReferenceEquals(field, option))
-                return;
-
-            field = option;
-            OnPropertyChanged(propertyName);
-        }
-
-        private string _effectiveEncryptionDisplay = string.Empty;
-        public string EffectiveEncryptionDisplay
-        {
-            get => _effectiveEncryptionDisplay;
-            set => SetField(ref _effectiveEncryptionDisplay, value ?? string.Empty, nameof(EffectiveEncryptionDisplay));
-        }
-
-        private bool _hasEncryptionSecret;
-        public bool HasEncryptionSecret
-        {
-            get => _hasEncryptionSecret;
-            set => SetField(ref _hasEncryptionSecret, value, nameof(HasEncryptionSecret));
-        }
-
-        private string _encryptionSecretStatus = string.Empty;
-        public string EncryptionSecretStatus
-        {
-            get => _encryptionSecretStatus;
-            set => SetField(ref _encryptionSecretStatus, value ?? string.Empty, nameof(EncryptionSecretStatus));
-        }
-
-        // Avatar
-        public string AvatarInitials { get; private set; } = string.Empty;
-        public string AvatarColor    { get; private set; } = "#33405A";
-        public string? AvatarImagePath { get; private set; }
-        public bool HasCustomAvatar => !string.IsNullOrWhiteSpace(AvatarImagePath);
-
-        public void SetAvatarFromNameAndStore(string name, string projectPath, string? externalId)
-        {
-            AvatarInitials  = ComputeInitials(name);
-            AvatarColor     = AvatarColorProvider.GetColor(name, projectPath, externalId);
-            AvatarImagePath = AvatarStore.GetAvatarForProject(projectPath);
-        }
-
-        private static string ComputeInitials(string name)
-        {
-            if (string.IsNullOrWhiteSpace(name))
-                return "?";
-
-            string[] parts = name.Split(new[] { ' ', '-', '_' }, StringSplitOptions.RemoveEmptyEntries);
-
-            if (parts.Length >= 2)
-                return $"{parts[0][0]}{parts[1][0]}".ToUpperInvariant();
-
-            if (name.Length >= 2)
-                return name.Substring(0, 2).ToUpperInvariant();
-
-            return name.Substring(0, 1).ToUpperInvariant();
-        }
-
-        public string LastBackupDisplay =>
-            LastBackupTime.HasValue
-                ? LastBackupTime.Value.ToString("yyyy-MM-dd HH:mm")
-                : LocalizationProvider.Service?.GetString("Backups.Summary.NoBackups") ?? "No backups yet";
-
-        public string TotalSizeFormatted =>
-            BackupSnapshotItem.FormatSize(TotalSizeBytes);
-
-        public bool HasStorageDelta => StorageDeltaBytes.HasValue;
-
-        public string StorageDeltaFormatted
-        {
-            get
-            {
-                if (!StorageDeltaBytes.HasValue)
-                    return "Δ -";
-
-                long value = StorageDeltaBytes.Value;
-                if (Math.Abs(value) < 1024)
-                    return "Δ ~0 B";
-
-                string sign = value >= 0 ? "+" : "-";
-                return $"Δ {sign}{BackupSnapshotItem.FormatSize(Math.Abs(value))}";
-            }
-        }
-
-        private void RebuildProjectTags()
-        {
-            ProjectTagChips.Clear();
-            foreach (ProjectTagChip chip in ProjectTagAppearance.CreateChips(_projectTagsCsv))
-                ProjectTagChips.Add(chip);
-        }
-    }
-
-    public class DestinationStatusItem : ViewModelBase
-    {
-        private static string L(string key, string fallback) =>
-            LocalizationProvider.Service?.GetString(key) ?? fallback;
-
-        private static readonly IBrush SuccessBrush = new ImmutableSolidColorBrush(Color.Parse("#22CC88"));
-        private static readonly IBrush WarningBrush = new ImmutableSolidColorBrush(Color.Parse("#FFB84C"));
-        private static readonly IBrush ErrorBrush = new ImmutableSolidColorBrush(Color.Parse("#FF6B6B"));
-        private static readonly IBrush InfoBrush = new ImmutableSolidColorBrush(Color.Parse("#8E9BAF"));
-
-        public string Id { get; set; } = string.Empty;
-        public string Alias { get; set; } = string.Empty;
-        public string Path { get; set; } = string.Empty;
-
-        private BackupsViewModel.DestinationStatus _status = BackupsViewModel.DestinationStatus.None;
-        public BackupsViewModel.DestinationStatus Status
-        {
-            get => _status;
-            set
-            {
-                if (_status == value)
-                    return;
-                _status = value;
-
-                OnPropertyChanged(nameof(Status));
-                OnPropertyChanged(nameof(StatusDisplay));
-                OnPropertyChanged(nameof(IsChecking));
-            }
-        }
-
-        public string StatusDisplay => _status switch
-        {
-            BackupsViewModel.DestinationStatus.Pending => L("Backups.Destinations.Pending", "Pending"),
-            BackupsViewModel.DestinationStatus.Inactive => L("Backups.Destinations.Inactive", "Inactive"),
-            BackupsViewModel.DestinationStatus.Reachable => L("Destinations.Test.Reachable", "Reachable"),
-            BackupsViewModel.DestinationStatus.ReadOnly => L("Destinations.Test.ReadOnly", "Read-only"),
-            BackupsViewModel.DestinationStatus.Unavailable => L("Destinations.Test.Unavailable", "Unavailable"),
-            BackupsViewModel.DestinationStatus.None => string.Empty,
-
-            _ => string.Empty
-        };
-        public void RefreshLocalization()
-        {
-            OnPropertyChanged(nameof(StatusDisplay));
-            OnPropertyChanged(nameof(LastCheckedDisplay));
-        }
-
-        public bool IsChecking => _status ==  BackupsViewModel.DestinationStatus.Pending;
-
-        private bool _isActive = true;
-        public bool IsActive
-        {
-            get => _isActive;
-            set => SetField(ref _isActive, value);
-        }
-
-        private bool _isConfigurable = true;
-        public bool IsConfigurable
-        {
-            get => _isConfigurable;
-            set => SetField(ref _isConfigurable, value);
-        }
-
-        private BackupsViewModel.SeverityStatus _severity = BackupsViewModel.SeverityStatus.None;
-        public BackupsViewModel.SeverityStatus Severity
-        {
-            get => _severity;
-            set
-            {
-                if (SetField(ref _severity, value))
-                {
-                    OnPropertyChanged(nameof(ReachabilityBrush));
-                }
-            }
-        }
-
-        private IBrush _dotBrush = InfoBrush;
-        public IBrush DotBrush
-        {
-            get => _dotBrush;
-            set => SetField(ref _dotBrush, value);
-        }
-
-        public IBrush ReachabilityBrush
-        {
-            get
-            {
-                return Severity switch
-                {
-                    BackupsViewModel.SeverityStatus.Success => SuccessBrush,
-                    BackupsViewModel.SeverityStatus.Warning => WarningBrush,
-                    BackupsViewModel.SeverityStatus.Error => ErrorBrush,
-                    BackupsViewModel.SeverityStatus.None => InfoBrush,
-                    _ => InfoBrush
-                };
-            }
-        }
-
-        private DateTime? _lastCheckedUtc;
-        public DateTime? LastCheckedUtc
-        {
-            get => _lastCheckedUtc;
-            set
-            {
-                if (SetField(ref _lastCheckedUtc, value))
-                {
-                    OnPropertyChanged(nameof(LastCheckedDisplay));
-                }
-            }
-        }
-
-        public string LastCheckedDisplay
-        {
-            get
-            {
-                if (!LastCheckedUtc.HasValue)
-                {
-                    return LocalizationProvider.Service?.GetString("Destinations.Status.LastCheckedNever")
-                           ?? "Last checked: never";
-                }
-
-                string label = LocalizationProvider.Service?.GetString("Destinations.Status.LastChecked")
-                           ?? "Last checked: {0}";
-                string local = LastCheckedUtc.Value.ToLocalTime().ToString("HH:mm:ss");
-                return string.Format(CultureInfo.CurrentCulture, label, local);
-            }
-        }
-
-        private string _storedBytesText = string.Empty;
-        public string StoredBytesText
-        {
-            get => _storedBytesText;
-            set
-            {
-                if (SetField(ref _storedBytesText, value))
-                {
-                    OnPropertyChanged(nameof(HasStoredBytesText));
-                }
-            }
-        }
-
-        public bool HasStoredBytesText => !string.IsNullOrWhiteSpace(StoredBytesText);
-
-        private string _cleanupSuggestionText = string.Empty;
-        public string CleanupSuggestionText
-        {
-            get => _cleanupSuggestionText;
-            set
-            {
-                if (SetField(ref _cleanupSuggestionText, value))
-                {
-                    OnPropertyChanged(nameof(HasCleanupSuggestionText));
-                }
-            }
-        }
-
-        public bool HasCleanupSuggestionText => !string.IsNullOrWhiteSpace(CleanupSuggestionText);
-
-        public static string GetId(BackupDestination dest) =>
-            DestinationIdentityService.GetId(dest);
-    }
-
-    public class BackupProgressItem : ViewModelBase
-    {
-        private static string L(string key, string fallback) =>
-            LocalizationProvider.Service?.GetString(key) ?? fallback;
-
-        public string ProjectId { get; set; } = string.Empty;
-        public string ProjectName { get; set; } = string.Empty;
-
-        public Action<BackupProgressItem?>? CancelRequested { get; set; }
-
-        private bool _allowCancel = true;
-        public bool AllowCancel
-        {
-            get => _allowCancel;
-            set
-            {
-                if (_allowCancel == value)
-                    return;
-
-                _allowCancel = value;
-                OnPropertyChanged(nameof(AllowCancel));
-                OnPropertyChanged(nameof(CanCancel));
-                OnPropertyChanged(nameof(IsIndeterminate));
-                NotifyProgressPresentationChanged();
-            }
-        }
-
-        public ICommand CancelCommand { get; }
-
-        public BackupProgressItem()
-        {
-            CancelCommand = new RelayCommand(_ => CancelRequested?.Invoke(this));
-        }
-
-        private string _destinationLabel = string.Empty;
-        public string DestinationLabel
-        {
-            get => _destinationLabel;
-            set
-            {
-                if (_destinationLabel == value)
-                    return;
-
-                _destinationLabel = value ?? string.Empty;
-                OnPropertyChanged(nameof(DestinationLabel));
-                OnPropertyChanged(nameof(DestinationDisplay));
-                OnPropertyChanged(nameof(HasDestinationDisplay));
-            }
-        }
-
-        public string DestinationDisplay
-        {
-            get
-            {
-                if (string.IsNullOrWhiteSpace(_destinationLabel))
-                    return string.Empty;
-
-                string prefix = L("Projects.List.DestinationPrefix", "Destination: ");
-                return $"{prefix}{_destinationLabel}";
-            }
-        }
-
-        public bool HasDestinationDisplay => !string.IsNullOrWhiteSpace(DestinationDisplay);
-
-        private string _policyText = string.Empty;
-        public string PolicyText
-        {
-            get => _policyText;
-            set
-            {
-                string normalized = value ?? string.Empty;
-                if (_policyText == normalized)
-                    return;
-
-                _policyText = normalized;
-                OnPropertyChanged(nameof(PolicyText));
-                OnPropertyChanged(nameof(HasPolicyText));
-            }
-        }
-
-        public bool HasPolicyText => !string.IsNullOrWhiteSpace(PolicyText);
-
-        private double _progress;
-        private double _displayProgress;
-        private string _lastStageKey = string.Empty;
-        private DateTime _stageStartUtc = DateTime.UtcNow;
-        public double Progress
-        {
-            get => _progress;
-            set
-            {
-                if (Math.Abs(_progress - value) < 0.0001)
-                    return;
-
-                _progress = value;
-                UpdateDisplayProgress();
-                OnPropertyChanged(nameof(Progress));
-                OnPropertyChanged(nameof(IsCompleted));
-                OnPropertyChanged(nameof(CanCancel));
-                OnPropertyChanged(nameof(IsEstimate));
-                NotifyProgressPresentationChanged();
-            }
-        }
-
-        public double DisplayProgress => _displayProgress;
-
-        public bool HasProgress => HasRawProgress;
-
-        private bool HasRawProgress => _progress > 0.1d;
-
-        private string _currentFile = string.Empty;
-        public string CurrentFile
-        {
-            get => _currentFile;
-            set
-            {
-                if (_currentFile == value)
-                    return;
-
-                _currentFile = value ?? string.Empty;
-                OnPropertyChanged(nameof(CurrentFile));
-                OnPropertyChanged(nameof(HasCurrentFile));
-                OnPropertyChanged(nameof(CurrentFileDisplay));
-                OnPropertyChanged(nameof(HasCurrentFileDisplay));
-                OnPropertyChanged(nameof(StageLabel));
-                OnPropertyChanged(nameof(IsEstimate));
-                UpdateDisplayProgress();
-                NotifyProgressPresentationChanged();
-            }
-        }
-
-        public bool HasCurrentFile => !string.IsNullOrWhiteSpace(_currentFile);
-
-        private string _etaText = string.Empty;
-        private string _lastProgressDetail = string.Empty;
-        public string EtaText
-        {
-            get => _etaText;
-            set
-            {
-                if (_etaText == value)
-                    return;
-
-                _etaText = value ?? string.Empty;
-                OnPropertyChanged(nameof(EtaText));
-                OnPropertyChanged(nameof(HasEtaText));
-                OnPropertyChanged(nameof(EtaDisplay));
-                OnPropertyChanged(nameof(HasEtaDisplay));
-                OnPropertyChanged(nameof(IsEstimate));
-                string detail = ExtractProgressDetail(EtaDisplay);
-                if (!string.IsNullOrWhiteSpace(detail))
-                {
-                    _lastProgressDetail = detail;
-                }
-                OnPropertyChanged(nameof(CurrentFileDisplay));
-                OnPropertyChanged(nameof(HasCurrentFileDisplay));
-                OnPropertyChanged(nameof(StageLabel));
-                UpdateDisplayProgress();
-                NotifyProgressPresentationChanged();
-            }
-        }
-
-        public bool HasEtaText => !string.IsNullOrWhiteSpace(_etaText);
-
-        public string EtaDisplay => NormalizeEtaText(_etaText);
-
-        public bool HasEtaDisplay => !string.IsNullOrWhiteSpace(EtaDisplay);
-
-        public bool IsEstimate =>
-            !HasRawProgress &&
-            ContainsToken(_currentFile, L("Backups.Progress.Estimating", "Estimating...")) &&
-            HasEtaText;
-
-        public static string EstimateLabel => L("Backups.Preflight.Title", "Backup estimate");
-
-        public string CurrentFileDisplay
-        {
-            get
-            {
-                if (TryExtractFileName(_currentFile, out string? fileName))
-                    return fileName;
-
-                if (!string.IsNullOrWhiteSpace(_lastProgressDetail) && !ContainsSpeedOrEta(_lastProgressDetail))
-                    return _lastProgressDetail;
-
-                return string.Empty;
-            }
-        }
-
-        public bool HasCurrentFileDisplay => !string.IsNullOrWhiteSpace(CurrentFileDisplay);
-
-        public bool IsCompleted => string.Equals(GetStageKey(), "Completed", StringComparison.OrdinalIgnoreCase);
-
-        public bool ShowEta => Progress < 100d && HasEtaDisplay;
-
-        public bool CanCancel => AllowCancel && !IsCompleted;
-
-        public bool ShowPercent => AllowCancel && HasProgress && IsProgressReliable;
-
-        public bool IsIndeterminate => !IsProgressReliable || IsStageIndeterminate;
-
-        public string ProgressLabel =>
-            IsProgressReliable && HasProgress
-                ? string.Format(CultureInfo.CurrentCulture, "{0:0}%", DisplayProgress)
-                : L("Backups.Progress.Estimating", "Estimating...");
-
-        public string StageLabel
-        {
-            get
-            {
-                string stageKey = GetStageKey();
-                return stageKey switch
-                {
-                    "Completed" => L("Backups.Status.Completed", "Completed"),
-                    "Cancelling" => L("Backups.Status.Cancelling", "Cancelling..."),
-                    "Deleting" => L("Backups.Stage.Deleting", "Deleting"),
-                    "Restoring" => L("Backups.Status.Restoring", "Restoring backup..."),
-                    "Compressing" => L("Backups.Stage.Compressing", "Compressing archive"),
-                    "Uploading" => L("Backups.Stage.Uploading", "Uploading archive"),
-                    "Hashing" => L("Backups.Stage.Hashing", "Hashing files"),
-                    "Copying" => L("Backups.Stage.Copying", "Copying files"),
-                    "Preparing" => L("Backups.Stage.Preparing", "Preparing"),
-                    "BackingUp" => L("Backups.Stage.BackingUp", "Backing up files"),
-                    _ => L("Backups.Stage.Working", "Working...")
-                };
-            }
-        }
-
-        public string StageDisplay
-            => IsCompleted ? StageLabel : $"{StageLabel} - {FormatElapsed(_stageStartUtc)}";
-
-        public IBrush StageBrush => GetStageBrush();
-
-        private bool IsProgressReliable
-        {
-            get
-            {
-                if (!HasRawProgress)
-                    return false;
-
-                return true;
-            }
-        }
-
-        private bool IsStageIndeterminate
-            => string.Equals(StageLabel, L("Backups.Stage.Preparing", "Preparing"), StringComparison.OrdinalIgnoreCase);
-
-        private bool IsHashingStage => string.Equals(GetStageKey(), "Hashing", StringComparison.OrdinalIgnoreCase);
-
-        private bool HasCompletionSignal()
-        {
-            if (ContainsToken(_etaText, "Completed"))
-                return true;
-
-            return ContainsToken(_currentFile, L("Backups.Status.Completed", "Completed"))
-                || ContainsToken(_currentFile, L("Backups.Status.NoChanges", "No changes detected"))
-                || ContainsToken(_currentFile, L("Backups.Status.Cancelled", "Cancelled"))
-                || ContainsToken(_currentFile, L("Backups.Status.Deleted", "Deleted"));
-        }
-
-        private static bool ContainsToken(string? value, string token)
-        {
-            if (string.IsNullOrWhiteSpace(value) || string.IsNullOrWhiteSpace(token))
-                return false;
-
-            return value.IndexOf(token, StringComparison.OrdinalIgnoreCase) >= 0;
-        }
-
-        private static string ExtractFileName(string value)
-        {
-            if (string.IsNullOrWhiteSpace(value))
-                return string.Empty;
-
-            string trimmed = value.Trim();
-
-            // Drop destination prefix like "[Alias]" if present.
-            if (trimmed.StartsWith("[", StringComparison.Ordinal))
-            {
-                int end = trimmed.IndexOf(']');
-                if (end >= 0 && end + 1 < trimmed.Length)
-                {
-                    trimmed = trimmed[(end + 1)..].Trim();
-                }
-            }
-
-            string candidate = trimmed;
-            if (candidate.Contains('\\') || candidate.Contains('/'))
-            {
-                candidate = Path.GetFileName(candidate);
-            }
-
-            return candidate;
-        }
-
-        private static bool TryExtractFileName(string value, out string fileName)
-        {
-            fileName = string.Empty;
-            if (string.IsNullOrWhiteSpace(value))
-                return false;
-
-            if (!value.Contains('\\') && !value.Contains('/'))
-                return false;
-
-            string extracted = ExtractFileName(value);
-            if (string.IsNullOrWhiteSpace(extracted))
-                return false;
-
-            fileName = extracted;
-            return true;
-        }
-
-        private static bool ContainsSpeedOrEta(string value)
-        {
-            if (string.IsNullOrWhiteSpace(value))
-                return false;
-
-            return value.Contains("ETA", StringComparison.OrdinalIgnoreCase)
-                   || value.Contains("MB/s", StringComparison.OrdinalIgnoreCase);
-        }
-
-        private static string ExtractProgressDetail(string value)
-        {
-            if (string.IsNullOrWhiteSpace(value))
-                return string.Empty;
-
-            string trimmed = value.Trim();
-            if (string.Equals(trimmed, L("Backups.Progress.CopyingRobocopy", "Copying files (robocopy)..."), StringComparison.OrdinalIgnoreCase))
-                return string.Empty;
-            if (trimmed.StartsWith("Copying ", StringComparison.OrdinalIgnoreCase))
-                return L("Backups.Progress.MovedPrefix", "moved ") + trimmed["Copying ".Length..];
-            if (trimmed.StartsWith("Compressing ", StringComparison.OrdinalIgnoreCase))
-                return trimmed["Compressing ".Length..];
-            if (trimmed.StartsWith("Uploading ", StringComparison.OrdinalIgnoreCase))
-                return trimmed["Uploading ".Length..];
-            if (trimmed.StartsWith("Hashing ", StringComparison.OrdinalIgnoreCase))
-                return trimmed;
-
-            return trimmed;
-        }
-
-        private static string NormalizeEtaText(string value)
-        {
-            if (string.IsNullOrWhiteSpace(value))
-                return string.Empty;
-
-            string trimmed = value.Trim();
-            if (trimmed.Contains("Waiting for first file", StringComparison.OrdinalIgnoreCase))
-                return L("Backups.Progress.WaitingForFirstFile", "Waiting for first file...");
-
-            if (trimmed.Contains("Copying files (robocopy)", StringComparison.OrdinalIgnoreCase))
-                return L("Backups.Progress.CopyingRobocopy", "Copying files (robocopy)...");
-
-            return trimmed;
-        }
-
-        private void NotifyProgressPresentationChanged()
-        {
-            OnPropertiesChanged(
-                nameof(DisplayProgress),
-                nameof(HasProgress),
-                nameof(IsIndeterminate),
-                nameof(ProgressLabel),
-                nameof(ShowPercent),
-                nameof(ShowEta),
-                nameof(StageLabel),
-                nameof(StageDisplay),
-                nameof(StageBrush),
-                nameof(EtaDisplay),
-                nameof(HasEtaDisplay));
-        }
-
-        private void UpdateDisplayProgress()
-        {
-            string stageKey = GetStageKey();
-            if (!string.Equals(stageKey, _lastStageKey, StringComparison.OrdinalIgnoreCase))
-            {
-                _displayProgress = 0d;
-                _lastStageKey = stageKey;
-                _stageStartUtc = DateTime.UtcNow;
-                OnPropertyChanged(nameof(StageDisplay));
-            }
-
-            if (!HasRawProgress || !IsProgressReliable)
-            {
-                _displayProgress = 0d;
-                return;
-            }
-
-            double next = Math.Clamp(_progress, 0d, 100d);
-            if (!HasCompletionSignal() && !IsHashingStage)
-            {
-                next = Math.Min(next, 99d);
-            }
-
-            if (IsCopyingStage && (DateTime.UtcNow - _stageStartUtc) < TimeSpan.FromSeconds(2) && next >= 99d)
-            {
-                next = Math.Min(next, 5d);
-            }
-
-            if (next < _displayProgress)
-            {
-                if (!HasCompletionSignal())
-                {
-                    _displayProgress = next;
-                }
-                return;
-            }
-
-            _displayProgress = next;
-        }
-
-        public void TickStageClock()
-        {
-            if (!IsCompleted)
-            {
-                OnPropertyChanged(nameof(StageDisplay));
-            }
-        }
-
-        private static string FormatElapsed(DateTime startUtc)
-        {
-            TimeSpan elapsed = DateTime.UtcNow - startUtc;
-            if (elapsed < TimeSpan.Zero)
-                elapsed = TimeSpan.Zero;
-
-            return elapsed.TotalHours >= 1
-                ? elapsed.ToString(@"hh\:mm\:ss", CultureInfo.InvariantCulture)
-                : elapsed.ToString(@"mm\:ss", CultureInfo.InvariantCulture);
-        }
-
-        private string GetStageKey()
-        {
-            if (ContainsToken(_currentFile, L("Backups.Status.Completed", "Completed")) ||
-                ContainsToken(_currentFile, L("Backups.Status.NoChanges", "No changes detected")) ||
-                ContainsToken(_currentFile, L("Backups.Status.Cancelled", "Cancelled")) ||
-                ContainsToken(_currentFile, L("Backups.Status.Deleted", "Deleted")) ||
-                (ContainsToken(_etaText, "Completed") && !ContainsToken(_etaText, "Hashing")))
-            {
-                return "Completed";
-            }
-
-            if (ContainsToken(_currentFile, L("Backups.Status.Cancelling", "Cancelling...")))
-                return "Cancelling";
-
-            if (ContainsToken(_etaText, "Compressing"))
-                return "Compressing";
-
-            if (ContainsToken(_etaText, "Uploading"))
-                return "Uploading";
-
-            if (ContainsToken(_etaText, "Hashing"))
-                return "Hashing";
-
-            if (ContainsToken(_etaText, "Copying"))
-                return "Copying";
-
-            if (ContainsToken(_etaText, "Restoring") ||
-                ContainsToken(_currentFile, "Restoring") ||
-                ContainsToken(_currentFile, "Decrypting") ||
-                ContainsToken(_currentFile, L("Backups.Status.Restoring", "Restoring backup...")))
-            {
-                return "Restoring";
-            }
-
-            if (ContainsToken(_currentFile, L("Backups.Status.Deleting", "Deleting backup files...")) ||
-                ContainsToken(_currentFile, L("Backups.Stage.Deleting", "Deleting")))
-            {
-                return "Deleting";
-            }
-
-            if (ContainsToken(_currentFile, L("Backups.Status.Preparing", "Preparing backup...")))
-                return "Preparing";
-
-            if (ContainsToken(_currentFile, "Reusing existing snapshot") ||
-                ContainsToken(_currentFile, "Creating snapshot") ||
-                ContainsToken(_currentFile, "Hashing"))
-            {
-                return "Hashing";
-            }
-
-            if (ContainsToken(_currentFile, L("Backups.Status.Running", "Running backup...")) ||
-                ContainsToken(_currentFile, L("Backups.Status.RunningMultiple", "Running backups...")))
-            {
-                return "BackingUp";
-            }
-
-            if (HasCurrentFile)
-                return "BackingUp";
-
-            return "Working";
-        }
-
-        private bool IsCopyingStage => string.Equals(GetStageKey(), "Copying", StringComparison.OrdinalIgnoreCase);
-
-        private static readonly IBrush StageCompletedBrush = new ImmutableSolidColorBrush(Color.Parse("#22CC88"));
-        private static readonly IBrush StageCancelBrush = new ImmutableSolidColorBrush(Color.Parse("#FF6B6B"));
-        private static readonly IBrush StageCompressBrush = new ImmutableSolidColorBrush(Color.Parse("#FFB84C"));
-        private static readonly IBrush StageUploadBrush = new ImmutableSolidColorBrush(Color.Parse("#22CCFF"));
-        private static readonly IBrush StageHashBrush = new ImmutableSolidColorBrush(Color.Parse("#9B6BFF"));
-        private static readonly IBrush StageCopyBrush = new ImmutableSolidColorBrush(Color.Parse("#4C8DFF"));
-        private static readonly IBrush StageBackupBrush = new ImmutableSolidColorBrush(Color.Parse("#3A7AFE"));
-        private static readonly IBrush StagePrepareBrush = new ImmutableSolidColorBrush(Color.Parse("#8E9BAF"));
-
-        private IBrush GetStageBrush()
-        {
-            string stageKey = GetStageKey();
-            return stageKey switch
-            {
-                "Completed"   => StageCompletedBrush,
-                "Cancelling"  => StageCancelBrush,
-                "Deleting"    => StageCancelBrush,
-                "Compressing" => StageCompressBrush,
-                "Uploading"   => StageUploadBrush,
-                "Hashing"     => StageHashBrush,
-                "Copying"     => StageCopyBrush,
-                "Restoring"   => StageCopyBrush,
-                "BackingUp"   => StageBackupBrush,
-                "Preparing"   => StagePrepareBrush,
-                _             => StagePrepareBrush
-            };
-        }
-
-    }
-
-    public class SnapshotActivityPoint
-    {
-        private static readonly IBrush SnapshotAutoBrush = new ImmutableSolidColorBrush(Color.Parse("#3A7AFE"));
-        private static readonly IBrush SnapshotManualBrush = new ImmutableSolidColorBrush(Color.Parse("#22CC88"));
-        private static readonly IBrush SnapshotImportedBrush = new ImmutableSolidColorBrush(Color.Parse("#FFB84C"));
-        private static readonly IBrush SnapshotEmptyBrush = new ImmutableSolidColorBrush(Color.Parse("#22FFFFFF"));
-        public string DayLabel { get; set; } = string.Empty;
-        public bool ShowLabel { get; set; } = true;
-        public int AutoCount { get; set; }
-        public int ManualCount { get; set; }
-        public int ImportedCount { get; set; }
-        public long TotalBytes { get; set; }
-        public double AutoHeight { get; set; }
-        public double ManualHeight { get; set; }
-        public double ImportedHeight { get; set; }
-        public bool IsEmpty => AutoCount + ManualCount + ImportedCount == 0;
-        public double EmptyHeight => IsEmpty ? 8 : 0;
-        public bool HasAuto => AutoCount > 0;
-        public bool HasManual => ManualCount > 0;
-        public bool HasImported => ImportedCount > 0;
-        public IBrush AutoBrush { get; set; } = SnapshotAutoBrush;
-        public IBrush ManualBrush { get; set; } = SnapshotManualBrush;
-        public IBrush ImportedBrush { get; set; } = SnapshotImportedBrush;
-        public IBrush EmptyBrush { get; set; } = SnapshotEmptyBrush;
-        public string TooltipText { get; set; } = string.Empty;
-    }
-
-    public sealed class RestoreReadinessIssueItem
-    {
-        public RestoreReadinessIssueItem(string projectName, string stateLabel, string reason, IBrush stateBrush)
-        {
-            ProjectName = projectName;
-            StateLabel = stateLabel;
-            Reason = reason;
-            StateBrush = stateBrush;
-        }
-
-        public string ProjectName { get; }
-        public string StateLabel { get; }
-        public string Reason { get; }
-        public IBrush StateBrush { get; }
-    }
-
 }
